+++ database/migrations/00003_create_schema_table.sql
@@ -3,7 +3,7 @@
 SELECT 'up SQL query';
 -- +goose StatementEnd
 create table if not exists Schema (
-    id                  integer primary key,
+    id                  BIGSERIAL primary key,
     name              varchar not null
 
 );
